Create a real Facade
It's unclear how, until now, the facade implementation we had has worked. Whatever the reason, it was failing when trying to do anything because it wasn't set up correctly as a real facade.

We now have Services\Sendportal to act as the real implementation for Facades\Sendportal, and we register the facade as we are supposed to, etc.

Additionally, some of the facade method names have been updated to be clearer about what they are doing. E.g. we change `currentWorkspaceIdResolver()` to `setCurrentWorkspaceIdResolver()`, which is more accurate as to its function.