Add tests and add loop for message deletion
This adds a test for the deletion of a subscriber's messages and any associated message failures. In adding this test, I realised that - because we were performing a mass deletion - eloquent events were not being fired. This meant that the failures were not being deleted and the operation would bomb out due to a constraint error. I'm in two minds as to whether the loop and delete per-message is better than a mass deletion with a loop that just handles the messages in our subscriber class. I think the deletion of a message's failures should be done at the Message level to ensure consistent behaviour and in practice, this should be a relatively inexpensive operation.