Self-host CSS and JS assets
Moving away from using CDNs for hosting assets, as there is no real
advantage to doing so for an application like SendPortal, while
having all the disadvantages of CDNs.

This also upgrades some of the versions to the latest available.

We have decided for now not to build the assets ourselves, as we
didn't want to take on the maintenance burden of a more complex
build process, without any immediately obvious benefit. However,
we should occasionally look to see if there are more up-to-date
versions to pull in, and perhaps can reconsider building it
ourselves in the future.